pyproject.toml
[tool.poetry]
name = "dojobot"
version = "0.1.0"
description = ""
authors = ["Daniel Pope <mauve@mauveweb.co.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"
pygithub = "^1.50"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
